+++ graspologic/layouts/__init__.py
@@ -3,12 +3,12 @@
 
 from graspologic.layouts.classes import NodePosition
 from graspologic.layouts.visualize import save_graph, show_graph
-from graspologic.layouts import layouts, render_only, tsne, nooverlap
+from graspologic.layouts import render_only, tsne, nooverlap
 from graspologic.layouts.layout_from_edges import (
     layout_node2vec_tsne_from_file,
     layout_node2vec_umap_from_file,
-    layout_with_node2vec_tsne,
     layout_with_node2vec_umap,
+    layout_with_node2vec_tsne,
     remove_overlaps,
 )
 
@@ -22,7 +22,7 @@ __all__ = [
     "nooverlap",
     "layout_node2vec_tsne_from_file",
     "layout_node2vec_umap_from_file",
-    "layout_with_node2vec_tsne",
     "layout_with_node2vec_umap",
+    "layout_with_node2vec_tsne",
     "remove_overlaps",
 ]
