+++ graspologic/layouts/nooverlap/__init__.py
@@ -2,10 +2,15 @@
 # Licensed under the MIT license.
 
 
-from graspologic.layouts.nooverlap.nooverlap_quad import quad_tree, node
+from graspologic.layouts.nooverlap.nooverlap_quad import (
+    quad_tree,
+    node,
+    remove_overlaps,
+)
 
 
 __all__ = [
     "quad_tree",
     "node",
+    "remove_overlaps",
 ]
